Adds logging for API login.
Now we can define a dictConfig value to log every attempt of login in DLRN API.
If no values for "API_LOG_CONFIG" and "API_LOG_PATH" is given during the wsgi creation at app.config variable, then the logs are sent to wsgi stream.

For doing that we are going to use flask_container_scaffold logging file.

Change-Id: Iafdca9aa1981109de1a94bf1f99dcc977bdb1e56