Remove -em and -eol tags when calculating the release
In puppet projects, DLRN uses the last tag to find out if it's building
a post-release commit. When last tag is a -em or -eol tag, the logic is
broken as it sees version <version>-em which is different to what
what is in metadata.json.

This patch is removing EM and EOL tags so that we can be sure that git
describe is not affected by them.

Change-Id: I48e3f3fb5f9085d85ee480b4fa57026a2b15d5b3