Set /var/tmp/pkgsrc as safe.directory in the mock
git has been updated in CS9 to 2.39 recently. This includes a check of
directory permissions to make sure that the owner of the repo is the
same running the git command [1].

This is breaking dlrn when running git commands inside the mock as the
effective user in the mock is different to the owner of the directory
(the dlrn builder user). This patch is setting /var/tmp/pkgsrc as
safe.directory to skip this check for this directory, which is where the
code is copied in the mock by dlrn.

[1] https://github.com/git/git/blob/v2.39.0/Documentation/RelNotes/2.31.4.txt

Change-Id: Id4b1a2f1f6261248882ca817729c72fe547e551a