Cancel brew task if timeout is excedeed.
Current timeout is set to 3600. If timeout is excedeed DLRN used to show error in the build despite that the task is running in brew and might be successful.
So with this, we take the running task id by using brew command and filtering the values with the koji_build_target and the src_rpm. So we avoid to take other running tasks because all dlrn builders uses the same user.

Change-Id: I08655b90e7c070403ff993a8e2ffef8e310702d4