Manage EOL tags automatically
When a project create EOL tag for a specific branch in openstack
project, the branch is removed. That makes DLRN to fallback to master if
the project is not pinned in rdoinfo which is an undesired behavior.

This patch is implementing an automatic mechanism to look for -eol tags
when DLRN fails to checkout to an stable upstream branch. The fallback
behavior implemented by this patch is:

- If dlrn fails to clone the repo to the instance default source branch,
  it will try to checkout <release>-eol tag. It it does not exist it
  try to fallback to master or main branch if fallback_to_master
  parameter is set to 1 (default value).
- If dlrn fails to clone branch named *-rdo and fallback_to_master is 1
  it will try to fallback to rpm-master.
- If dlrn fails to clone a branch which is not the default source or
  ends with -rdo it will fail.

Change-Id: I2726321b7d64a28851582400e8af64d020396f48