Switch antelope src branch to release date-based name
Currently, DLRN-rpmbuild-base jobs for antelope are
falling back to master branch as the source branch
'stable/antelope' does not exist. Upstream, moved to
release date-based name.
We need to adapt our job to checkout to the right
branch.
Currently, we don't have such metadata in rdoinfo,
so I'm adding support to Antelope in order to
unblock CI job, but we need to handle it dynamically.

Change-Id: I60d6b574b416f811277b3f9b8bc58f33cd6b2772