Tweak DLRN tests to unpin sh version.
Last sh version for python3.6 is 1.14.3. We have pinned it to 1.14.2 because of the tests started to fail.
Zuul works with python3.9 that has version 2.0.3 from sh library, so we don't want to take that major version.
After mocking sh execution diferently, we are able to unpin the version to take the latest.

Change-Id: I05074939c42cd12ded6f45acd7dd40fcb0b284d6