Support components removal when creating aggregated repo files
Currently, in a dlrn instance with components enabled, when creating
aggregated repo files (delorean.repo and versions.csv under /current
directory, i.e.), the list of the components is taken as the list of
components of all the builds in the database.

This is not working when we need to remove a component from a running
dlrn instance, as we are currently doing with tripleo component [1].
Actually we are getting the last status of tripleo component even when
we don't have any package assigned to it in rdoinfo.

This patch is creating that list from the intersection of the whole list
of components in builds in db and the list of components assigned to
packages in the built tag. That should fix the components removal and
also the bootstraping situation, where some components are still empty.

As effect of this patch, the dlrn api is now using the pkginfo driver
which were not using until now.

[1] https://review.rdoproject.org/r/c/rdoinfo/+/47883

Change-Id: I8e06bfc7b264d70df5c250ac9b766587fba357a9