Order components when aggregating delorean.repo file
The md5 checksum of the aggregated delorean.repo files is used by CI as
a mean of repo identifier so we need to make it predictable to make sure
that two aggregated repos with the same components repos have the same
md5.

This patch is ordering the list of components when aggregating the
individual delorean.repo files into a ingle aggregated delorean.repo.

Change-Id: Ib3daceec3b7ef64012ed07f782cfc3cec271c321