Move dlrn-api-functional-centos-8 to build for centos8-yoga
Currently it's building packages for centos8-master which is not a
supported combination (master branches are only supported and tested for
centos9.

This patch is moving the dlrn-api-functional-centos-8 job to build
packages for yoga release on centos8 buildroot. Yoga is the last release
supported on centos8 so should be a good reference to validate dlrn.

Note that, we will be adding a new job dlrn-api-functional-centos-9
which will run on centos9 and will build packages for centos9 in one of
its supported releases.

Change-Id: I99a76f2ea3cc37380f40b5bb8e0387f2b30aeaca