Add Role system for Kerberos driver.
Add role system for kerberos driver, instead of checking if the given user in the kerberos token is member of group defined with "ALLOWED_GROUP", now it returns all the groups that the given user is membership and checks if any of those groups are the specified by: "API_READ_WRITE_ROLES" and/or "API_READ_ONLY_ROLES".

"API_READ_ONLY" is set for GET HTTP methods endpoints, and that role is subset of "API_READ_WRITE_ROLES" that is set for POST HTTP methods.

Change-Id: Ia391037c0e55af2315ec23508867b32e4598a41b