Implement decorator for retry method execution
Allows use of a decorator for retrying a number of times the execution of the decorated method.
With this, we avoid an unauthorized request while using authorization with IPA because of transient network errors or IPA unexpected errors.

This is a prerequisite for removing authorization from authentication drivers, so in a following patch we will create an authorization class that defines an interface for abstraction in authorization drivers.
Those drivers would use the connection_decorator that receives an action/start message, success/finish message and a custom error that is a list of Errors to retry if raised.

If any other error is raised, then it is an unexpected problem or a ConfigError that might produced a Failure authorization.

Change-Id: Ib26b0c99db210357c93d1078b17d3441a3f4aaac