Failback to unmaintained/<release> branch if exists
OpenStack is moving to a new model for releases after maintained phase
finishes [1].

The summary is that Extended Maintenance is replaced by a new
Unmaintained phase with a better defined and shorter lifecycle and only
SLURP releases are eligible (with exceptions for the current EM
releases).

Part of this change is that, once maintained phase is finished,
stable/<release> branches are removed and new branches
unmaintained/<release> are created.

[1] https://review.opendev.org/c/openstack/governance/+/888771

Depends-On: https://softwarefactory-project.io/r/c/DLRN/+/29166
Change-Id: I8c4710c30c341ab55a56cba3b372b3c706fc034b