Add ability to keep VERSION set in SPEC file of python pkgs
If we want to rebuild a package which already have Version
set in SPEC file, we may want to reuse same Version value and
avoid getting the version with the "python setup.py --version"
command again. This patch adds the possibility to keep same
Version value by setting DLRN_KEEP_VERSION to 1.

Change-Id: Ibdf2d0ee36b16cbf31525d871a62e0dd53e32783