Adding NVR and commit hash to status_report jinja2
This is using an already created utils method for taking the nvr from a
srpm list [1].
pkg.artifacts is a list of packages where the first one is the srpm,
so here we need to split the pkg value and take the first element. From
here we can use the utils method by converting that element in a list.

[1]: https://github.com/softwarefactory-project/DLRN/blob/master/dlrn/utils.py#L241

Change-Id: I045bcaef453478fde247e4dbf9d285d33598e4e5