Remove remote-tracking ref that no longer exist
This patch enables the transition from stable to
unmaintained branches automatically.
When upstream project move to unmaintained
status, it removes the associated stable branch [1].
But, as the stable branch still exists locally, DLRN
checkout to it.
With this change, DLRN will sync branches removal from
remote to local. As the sable branch won't exist anymore,
it will try to locate unmaintained branch [2].

Note that we added the whole block inside the try-except block, so if
something goes wrong we remove the repo folder, and it will be cloned
again during the next DLRN run.

[1] https://docs.openstack.org/project-team-guide/stable-branches.html#unmaintained
[2] https://github.com/softwarefactory-project/DLRN/blob/debb84c364e1544ff96e7babafecc91f716f1126/dlrn/repositories.py#L97

Closes https://issues.redhat.com/browse/RDO-427

Change-Id: I9e502f89a0a80d5488f333b1db91b3c6af73bf9f