Fix consistency between dlrn/repositories and kojidriver
We'were taking the last commit from HEAD instead of "." in
DLRN repositories.

In the DownstreamInfoDriver, we were looking in the git.log but
it hasn't been changed from "." to HEAD. So empty last commits
we're taking the previous to HEAD so the info of what has been
built and what's next had a mismatch in commits ID but not in content.
This cause a infinite loop of building package that are already built.

Change-Id: Ibc0126aaafc6ec597f86445579d04a4219a35958